// set-2/ushur-ui/src/features/canvas/interfaces/module-types.ts
enum ModuleTypes {
  MESSAGE_MODULE = "message-module",
  MENU_MODULE = "menu-module",
  FORM_MODULE = "form-module",
  COMPUTE_MODULE = "compute-module",
  BRANCH_MODULE = "branch-module",
  AI_MODULE = "ai-module",
  CONFIGURED_MODULE = "configured-module",
}

export default ModuleTypes;